refactor: migrate CheckBox folder to typescript
(cherry picked from commit 8b706ecd1b9eb3f32c551bb70eba379cd76794bd)
